Logitech Flight Radio Panel
This rule so far has allowed me to use my Logitech Flight Radio Panel as a joystick rather than keyboard and mouse (wtf Logitech?). Got the buttons working so far under evdev. Still working on the js - I'm a bit of a novice at programming and learning as I go.